requirements.txt
streamlit==1.35.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
camelot-py==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.30.0
numpy==2.2.3
openai==1.84.0
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMuPDF==1.26.0
pyparsing==3.2.1
pypdf==5.7.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
pywin32-ctypes==0.2.3
pyzbar==0.1.9
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.43.2
tabulate==0.9.0
tenacity==9.0.0
thefuzz==0.22.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
xlsxwriter==3.2.5
